+++ OversamplingOscillators.hpp
@@ -4,6 +4,7 @@
 
 #include "SC_PlugIn.hpp"
 #include "VariableOversampling.hpp"
+#include <array>
 
 namespace SawOS {
 
@@ -88,7 +89,7 @@ class SinOscOS : public SCUnit {
 
     float *osBuffer;
 
-    double sample_rate;
+    float sample_rate;
     float m_freqMul{2.0f/(float)sampleRate()};
     int m_oversamplingIndex{(int)in0(OverSample)};
   };
@@ -116,7 +117,7 @@ class PMOscOS : public SCUnit {
     enum InputParams { CarFreq, ModFreq, PMMul, PMModPhase, OverSample, NumInputParams };
     enum Outputs { Out1, NumOutputParams };
 
-    double sample_rate{(float)sampleRate()};
+    float sample_rate{(float)sampleRate()};
     float m_modphase{in0(PMModPhase)};
     float m_freqMul{2.0f/(float)sampleRate()};
     int m_oversamplingIndex{(int)in0(OverSample)};
@@ -144,7 +145,7 @@ class FM7OS : public SCUnit {
     OverSample, NumInputParams };
     enum Outputs { Out1, Out2, Out3, Out4, Out5, Out6, NumOutputParams };
 
-    double sample_rate{(float)sampleRate()};
+    float sample_rate{(float)sampleRate()};
     
     float m_phases[6] = {in0(ctl1),in0(ctl4),in0(ctl7),in0(ctl0),in0(ctl3),in0(ctl6)};
 
@@ -170,14 +171,7 @@ class FM7aOS : public SCUnit {
     // Calc function
     void next_aa(int nSamples);
 
-    //6, 16, 
-    // enum InputParams { ctl0, ctl1, ctl2, ctl3, ctl4, ctl5, ctl6, ctl7, ctl8, ctl9, ctl10, ctl11, ctl12, ctl13, ctl14, ctl15, ctl16, ctl17, 
-    // modNum0, modNum1, modNum2, modNum3, modNum4, modNum5, modNum6, modNum7, modNum8, modNum9, modNum10, modNum11, modNum12, modNum13, modNum14, modNum15, modNum16, modNum17, modNum18, modNum19, modNum20, modNum21, modNum22, modNum23, modNum24, modNum25, modNum26, modNum27, modNum28, modNum29, modNum30, modNum31, modNum32, modNum33, modNum34, modNum35,
-    // //oscType0, oscType1, oscType2, oscType3, oscType4, oscType15,
-    // OverSample, NumInputParams };
-    // enum Outputs { Out1, Out2, Out3, Out4, Out5, Out6, NumOutputParams };
-
-    double sample_rate{(float)sampleRate()};
+    float sample_rate{(float)sampleRate()};
     
     float m_phases[4] = {0.f, 0.f,0.f, 0.f};
 
@@ -205,7 +199,7 @@ class FM7bOS : public SCUnit {
     // Calc function
     void next_aa(int nSamples);
 
-    double sample_rate{(float)sampleRate()};
+    float sample_rate{(float)sampleRate()};
     
     float m_phases[4] = {0.f, 0.f,0.f, 0.f};
 
@@ -240,7 +234,7 @@ class PM7OS : public SCUnit {
     OverSample, NumInputParams };
     enum Outputs { Out1, Out2, Out3, Out4, Out5, Out6, NumOutputParams };
 
-    double sample_rate{(float)sampleRate()};
+    float sample_rate{(float)sampleRate()};
     
     float m_phases[6] = {in0(ctl1),in0(ctl4),in0(ctl7),in0(ctl0),in0(ctl3),in0(ctl6)};
 
@@ -376,7 +370,7 @@ public:
   // Destructor
   ~SergeFoldOS();
   VariableOversampling<> oversample;
-  std::array<float, 1000> sergeWavetable;
+  std::array<double, 1000> sergeWavetable;
 
 private:
   // Calc function
